requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

# ML
ultralytics==8.3.10
opencv-python-headless==4.10.0.84
torch==2.2.0
numpy==1.26.4

# Database
psycopg2-binary==2.9.9
requests==2.32.2

# Backend
fastapi==0.115.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
websockets==12.0
python-dotenv==1.0.1
